Not sure val_loss but it's pretty good

+++ train_eye/convert_images.py
@@ -10,7 +10,7 @@ if __name__ == "__main__":
         pass
     for img_file in imgs_files:
         try:
-            if not(os.path.exists(("images/{0}".format(img_file)))):
+            if not(os.path.exists(("images/{0}".format(img_file))) and not os.path.exists(("images/{0}".format(img_file)))):
                 img = cv2.imread("images/{0}".format(img_file))
                 img = cnn.img_reprocess(img, crop_size=250, img_size=224, to_greyscale=False)
                 cv2.imwrite("output_VGG/{0}".format(img_file),img)
